Store key material using crypto-pubkey-types
This is motivated by the fact that RFC6637 has key material that is not
just a sequence of MPIs.  crypto-pubkey-types is a pretty lightweight
dependency and has types for many of our keys (except elgamal).  This
also gets rid of the problem where you could have a piece of data with a
key algorithm specified but not the expected key fields for that
algorithm, etc, removing a possible runtime error case.

Conveniently, this also allows us to get rid of the icky encrypted_data
field on secret keys, since this just becomes a different kind of key
material.

Several helpers are introduced, including key_packet_algorithm, which
extracts the KeyAlgorithm from the relevant key material, and
public_material, which extracts the PublicKeyMaterial from both
PublicKeyPacket and SecretKeyPacket.